Drops watchOS + tvOS support (*)
Thanks Couchbase.
Thanks Cocoapods.

Podspecs support per-subpec deployment targets, but linting fails when doing this.
No one's gonna be using this on watchOS and probably not tvOS, so it's probably not a big deal.
A workaround would be to create separate podspecs, one that includes CouchbaseLite and one that does not.